Moved deserialization to its own class
