+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/CreateMeteringPoint/Input/CreateMeteringPointXmlDeserializer.cs
@@ -1,0 +1,80 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Xml.Linq;
+using Energinet.DataHub.MeteringPoints.Application;
+
+namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.CreateMeteringPoint.Input
+{
+    public static class CreateMeteringPointXmlDeserializer
+    {
+        public static IEnumerable<Application.CreateMeteringPoint> Deserialize(string bodyStr)
+        {
+            var root = XElement.Parse(bodyStr);
+            XNamespace ns = "urn:ediel:org:requestchangeofapcharacteristics:0:1";
+
+            var marketActivityRecords = root
+                .Elements(ns + "MktActivityRecord");
+
+            return marketActivityRecords.Select(record =>
+            {
+                var marketEvaluationPoint = record.Element(ns + "MarketEvaluationPoint");
+                var mainAddress = marketEvaluationPoint?.Element(ns + "usagePointLocation.mainAddress");
+                var streetDetail = mainAddress?.Element(ns + "streetDetail");
+                var townDetail = mainAddress?.Element(ns + "townDetail");
+                var contractedConnectionCapacity =
+                    marketEvaluationPoint?.Element(ns + "marketAgreement.contractedConnectionCapacity"); // TODO: MarketAgreement prefix isn't 100% set in stone. Update at a later point.
+                var series = marketEvaluationPoint?.Element(ns + "Series");
+
+                // Power plant
+                var linkedMarketEvaluationPoint = marketEvaluationPoint?.Element(ns + "Linked_MarketEvaluationPoint");
+
+                var address = new Address(
+                    ExtractElementValue(streetDetail, ns + "name"),
+                    ExtractElementValue(mainAddress, ns + "postalCode"),
+                    ExtractElementValue(townDetail, ns + "name"),
+                    ExtractElementValue(townDetail, ns + "country"),
+                    ExtractElementValue(mainAddress, ns + "usagePointLocation.remark") == "D01");
+
+                return new Application.CreateMeteringPoint(
+                    address,
+                    ExtractElementValue(marketEvaluationPoint, ns + "mRID"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "type"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "meteringMethod"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "readCycle"),
+                    Convert.ToInt32(ExtractElementValue(marketEvaluationPoint, ns + "ratedCurrent")),
+                    Convert.ToInt32(ExtractElementValue(contractedConnectionCapacity, ns + "value")),
+                    ExtractElementValue(marketEvaluationPoint, ns + "meteringGridArea_Domain.mRID"),
+                    ExtractElementValue(linkedMarketEvaluationPoint, ns + "mRID"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "usagePointLocation.remark"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "product"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "parent_MarketEvaluationPoint.mRID"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "settlementMethod"),
+                    ExtractElementValue(series, ns + "quantity_Measure_Unit.name"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "disconnectionMethod"),
+                    ExtractElementValue(record, ns + "start_DateAndOrTime.dateTime"),
+                    ExtractElementValue(marketEvaluationPoint, ns + "meter.mRID"),
+                    ExtractElementValue(record, ns + "mRID"));
+            });
+        }
+
+        private static string ExtractElementValue(XElement? element, XName name)
+        {
+            return element?.Element(name)?.Value ?? string.Empty;
+        }
+    }
+}
